# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "NeuroSuitUp API",
        "description": "Tools for communicating with the NeuroSuitUp database",
        "version": "1.0.0",
        "contact": {
        "name": "Apostolos Praftsiotis",
        "email": "apospraf@gmail.com"
        }
    },
    "servers": [
        {
            "url":"http://192.168.100.14:3000",
            "description":"Olpal server"
        },
        {
            "url":"https://locahost:3000",
            "description":"Temp localhost server"
        },
        {
            "url": "https://neurosuituptest.herokuapp.com/",
            "description": "Heroku free server"
        }
    ],
    "components": {
      "schemas": { 
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }, 
        "Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
      },
      "parameters": {
          "selectstring": {
              "in": "header",
              "name": "selectstring",
              "description": "Selects the fields of the returned resource",
              "schema": {
                  "type": "string"
              },
              "allowEmptyValue": true
          }
      },
      "responses": {
...
        "NotFound": {
            "description": "The specified resource was not found",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error" 
                    }
                }
            }
        }, 
        "Unauthorized": {
            "description": "Authorization failed, either by unsuccessfull verification of token, or by limited permissions",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }, 
        "InternalError": {
            "description": "Internal or structural error",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
      "/users/commandingInstructor/{patientId}": {
        "get": {
          "summary": "Get the commanding instructor of the patient with _id = patientId",
          "tags": ["users"],
          "parameters": [
            {
              "in": "path",
              "name": "patientId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "ObjectId (string) of the patient's commanding instructor"
            }
          ],
          "responses": {
            "200": {
              "description": "Found commanding instructor",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "401": {
                "$ref": "#/components/responses/Unauthorized"
            },
            "404": {
                "$ref": "#/components/responses/NotFound"
            },
            "500": {
                "$ref": "#/components/responses/InternalError"
            }
          }
        }
      },
      "/users/byRole/{role}": {
        "get": {
          "summary": "Get users with role = {role}",
          "tags": ["users"],
          "parameters": [
            {
              "in": "path",
              "name": "role",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "Requested role of users"
            },
            {
                "$ref": "#/components/parameters/selectstring"
            }
          ],
          "responses": {
            "200": {
              "description": "Query was succesfull, return the users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "count": {
                        "type": "integer",
                        "description": "Number of users found"
                      },
                      "users": {
                        "type": "array",
                        "description": "The users that were found",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
                "description": "Bad request, probably the role that was requested does not exist",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Message"
                        }
                    }
                }
            },
            "401": {
                "$ref": "#/components/responses/Unauthorized"
            },
            "500": {
                "$ref": "#/components/responses/InternalError"
            }
          }
        }
      },
      "/users/self/{userId}": {
        "get": {
            "summary": "Get information about the logged in user. Accessible by self only",
            "tags": ["users"],
            "parameters": [{
                "in": "path",
                "name": "userId",
                "schema": {
                    "type": "string"
                },
                "required": true,
                "description": "The user id (ObjectId, mongo)"
            }],
            "responses": {
                "200": {
                    "description": "Query was succesfull, return the user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "404": {
                    "$ref": "#/components/responses/NotFound"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/users/byId/{userId}": {
          "get": {
              "summary": "Get user information by id",
              "tags": ["users"],
              "parameters":[{
                  "in": "path",
                  "name": "userId",
                  "schema": {
                      "type": "string"
                  },
                  "required": true,
                  "description": "The id of the requested user"
              }, 
              {
                "$ref": "#/components/parameters/selectstring"
              }], 
              "responses": {
                  "200": {
                      "description": "Query was succesfull, return the user",
                      "content": {
                          "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/User"
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "404": {
                      "$ref": "#/components/responses/NotFound"
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      },
      "/users/tokenInfo/{token}": {
          "get": {
              "summary": "Get the information stored in the given token",
              "tags": ["users"],
              "parameters": [{
                  "in": "path",
                  "name": "token",
                  "schema": {
                      "type": "string"
                  },
                  "required": true,
                  "description": "The requested token"
              }],
              "responses": {
                  "200": {
                      "description": "Query was succesfull, return the token info (payload)",
                      "content": {
                          "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      }, 
      "/users/admin/{userId}": {
        "get": {
            "summary": "Get user information by id with admin authorization",
            "tags": ["users"],
            "parameters": [{
                "in": "path",
                "name": "userId",
                "schema": {
                    "type": "string"
                },
                "required": true,
                "description": "The requested user id"
            },
            {
                "$ref": "#/components/parameters/selectstring"
            }],
            "responses": {
                "200": {
                    "description": "Query was succesfull, return the user",
                    "content": {
                        "application/json": {
                            "schema": {
                              "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "404": {
                    "$ref": "#/components/responses/NotFound"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/admin": {
        "get": {
            "summary": "Get all users, accessible only by admins",
            "tags": ["users"],
            "parameters": [
                {
                    "$ref": "#/components/parameters/selectstring"
                }
            ],
            "responses": {
                "200": {
                  "description": "Query was succesfull, return the users",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "count": {
                            "type": "integer",
                            "description": "Number of users found"
                          },
                          "users": {
                            "type": "array",
                            "description": "The users that were found",
                            "items": {
                                "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
...
      "/users/signup": {
        "post": {
            "summary": "Sign up new user",
            "tags": ["users"],
            "requestBody": {
                "description": "The user to sign up",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/UserInsert",
                            "required": ["username", "first_name", "last_name", "email", "password"]
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "Succefully created user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            },
                            "example": {
                                "message": "User created"
                            }
                        }
                    }
                },
                "401": {
                    "description": "Admin creation from an not admin user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "409": {
                    "description": "Username or email exists",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/responses/SimpleMessage"
                            },
                            "example": {
                                "message": "Username or email exists"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
               
        }
      },
      "/users/login": {
        "post": {
            "summary": "Login user with credentials.",
            "tags": ["users"],
            "requestBody": {
                "description": "The credentials",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "rUsername": {
                                    "type": "string"
                                },
                                "rPassword": {
                                    "type": "string"
                                }
                            },
                            "required": ["rUsername", "rPassword"]
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "User succefully logged in",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "User information and the access token",
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/User"
                                    },
                                    "token": {
                                        "type": "string",
                                        "description": "The access token"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "example": {
                                    "user": {
                                        "analytics": null,
                                        "_id": "615c94273b3ffd5ccc3db1dd",
                                        "first_name": "Loukas",
                                        "last_name": "Ouzounis",
                                        "username": "ouzos",
                                        "email": "ouzos@tsipour.aki",
                                        "hash": null,
                                        "age": 36,
                                        "gender": "Male",
                                        "history": "",
                                        "role": "instructor",
                                        "patients": [
                                            "61656f23daf470ec8561ae9d",
                                            "61656f6edaf470ec8561ae9e"
                                        ],
                                        "poses": [
                                            "6162e6e8f7e7cb5b5924287e",
                                            "615c94853b3ffd5cccec496c",
                                            "615c96a63b3ffd5cccdf8816",
                                            "615c978c3b3ffd5ccc72ddda",
                                            "61751fc6aa7ddca06ae996c5",
                                            "61751fe0aa7ddca06ae996ce"
                                        ],
                                        "tasks": [
                                            "61712f1f17c8a14be8d94fe6",
                                            "617131a5bcdd0d26dceaf1e0",
                                            "617131f7bcdd0d26dceaf202",
                                            "61713220bcdd0d26dceaf21d",
                                            "6172aa4c6d94f01e891f8f24",
                                            "61751feaaa7ddca06ae996d5",
                                            "6175205aaa7ddca06ae9972e"
                                        ],
                                        "exercises": [
                                            "6174378d6d01fd9aa345702a",
                                            "61743e69e7ec886742092075",
                                            "6172ab406d94f01e891f8f8b",
                                            "61751a6933fa0f240b8ec018",
                                            "61752002aa7ddca06ae996e3",
                                            "61752072aa7ddca06ae9973c"
                                        ],
                                        "viewed": true,
                                        "prescription": []
                                    },
                                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MTVjOTQyNzNiM2ZmZDVjY2MzZGIxZGQiLCJyb2xlIjoiaW5zdHJ1Y3RvciIsInVzZXJuYW1lIjoib3V6b3MiLCJpYXQiOjE2MzUxNjAxODksImV4cCI6MTYzNTE2Mzc4OX0.3oHbP7Y84wpMC3cB4DO2gmjkiQFt5D9W6Z0afEcTy8s",
                                    "messsage": "Auth successful"
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/users/assign": {
        "patch": {
            "summary": "Assign payient to intructor",
            "tags": ["users"],
            "requestBody": {
                "description": "The patientId",
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "type": "object", 
                            "properties": {
                                "patientId": {
                                    "type": "string"
                                }
                            },
                            "required": ["patientId"]
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "Patient successfully added to instructor list",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message", 
                                "example": {
                                    "patientId": "616092b63d00a77c869ba529"
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "409": {
                    "description": "Patient is already assigned to another instructor",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      }, 
      "/users/unassign": {
        "patch": {
            "summary": "Remove patient from logged intructor patient list",
            "tags": ["users"],
            "requestBody": {
                "description": "The patientId",
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "type": "object", 
                            "properties": {
                                "patientId": {
                                    "type": "string"
                                }
                            },
                            "required": ["patientId"]
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Patient successfully removed from instructor patient list",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "403": {
                    "description": "Patient is not assigned to the logged in instructor",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
...
    }
}